Implement OpsMixin
